requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain==0.0.327
chromadb==0.4.13
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
unstructured==0.10.28
HuggingFaceHub==0.17.3